+++ src/experimental/env.rs
@@ -116,7 +116,12 @@ mod sys {
             }
         }
     }
-    #[cfg(any(target_arch = "mips", target_arch = "mips64"))]
+    #[cfg(any(
+        target_arch = "mips",
+        target_arch = "mips32r6",
+        target_arch = "mips64",
+        target_arch = "mips64r6",
+    ))]
     mod imp {
 
         use core::cell::Cell;