+++ src/experimental/time.rs
@@ -210,7 +210,12 @@ mod sys {
             }
         }
     }
-    #[cfg(any(target_arch = "mips", target_arch = "mips64"))]
+    #[cfg(any(
+        target_arch = "mips",
+        target_arch = "mips32r6",
+        target_arch = "mips64",
+        target_arch = "mips64r6",
+    ))]
     mod inner {
         use super::SystemTime;
         use crate::io;