+++ src/sys/errno.rs
@@ -10,14 +10,24 @@ pub(crate) fn decode_error_kind(errno: RawOsError) -> io::ErrorKind {
     #[allow(clippy::enum_glob_use)]
     use io::ErrorKind::*;
     match errno {
-        #[cfg(not(any(target_arch = "mips", target_arch = "mips64")))] // TODO
+        #[cfg(not(any(
+            target_arch = "mips",
+            target_arch = "mips32r6",
+            target_arch = "mips64",
+            target_arch = "mips64r6",
+        )))] // TODO
         errno::E2BIG => __ArgumentListTooLong, // unstable
         // errno::EADDRINUSE => AddrInUse,
         // errno::EADDRNOTAVAIL => AddrNotAvailable,
         errno::EBUSY => __ResourceBusy, // unstable
         // errno::ECONNABORTED => ConnectionAborted,
         // errno::ECONNREFUSED => ConnectionRefused,
-        #[cfg(any(target_arch = "mips", target_arch = "mips64"))] // TODO
+        #[cfg(any(
+            target_arch = "mips",
+            target_arch = "mips32r6",
+            target_arch = "mips64",
+            target_arch = "mips64r6",
+        ))] // TODO
         errno::ECONNRESET => ConnectionReset,
         // errno::EDEADLK => Deadlock, // unstable
         // errno::EDQUOT => FilesystemQuotaExceeded, // unstable
@@ -36,7 +46,12 @@ pub(crate) fn decode_error_kind(errno: RawOsError) -> io::ErrorKind {
         // errno::ENAMETOOLONG => InvalidFilename, // unstable
         // errno::ENETDOWN => NetworkDown, // unstable
         // errno::ENETUNREACH => NetworkUnreachable, // unstable
-        #[cfg(any(target_arch = "mips", target_arch = "mips64"))] // TODO
+        #[cfg(any(
+            target_arch = "mips",
+            target_arch = "mips32r6",
+            target_arch = "mips64",
+            target_arch = "mips64r6",
+        ))] // TODO
         errno::ENOTCONN => NotConnected,
         errno::ENOTDIR => __NotADirectory, // unstable
         // errno::ENOTEMPTY => DirectoryNotEmpty, // unstable
@@ -44,7 +59,12 @@ pub(crate) fn decode_error_kind(errno: RawOsError) -> io::ErrorKind {
         errno::EROFS => __ReadOnlyFilesystem, // unstable
         errno::ESPIPE => __NotSeekable,       // unstable
         // errno::ESTALE => StaleNetworkFileHandle, // unstable
-        #[cfg(any(target_arch = "mips", target_arch = "mips64"))] // TODO
+        #[cfg(any(
+            target_arch = "mips",
+            target_arch = "mips32r6",
+            target_arch = "mips64",
+            target_arch = "mips64r6",
+        ))] // TODO
         errno::ETIMEDOUT => TimedOut,
         // errno::ETXTBSY => ExecutableFileBusy, // unstable
         errno::EXDEV => __CrossesDevices, // unstable
@@ -53,7 +73,12 @@ pub(crate) fn decode_error_kind(errno: RawOsError) -> io::ErrorKind {
         // These two constants can have the same value on some systems,
         // but different values on others, so we can't use a match
         // clause
-        #[cfg(any(target_arch = "mips", target_arch = "mips64"))] // TODO
+        #[cfg(any(
+            target_arch = "mips",
+            target_arch = "mips32r6",
+            target_arch = "mips64",
+            target_arch = "mips64r6",
+        ))] // TODO
         x if x == errno::EAGAIN || x == errno::EWOULDBLOCK => WouldBlock,
         _ => Other,
     }