+++ tools/no-std.sh
@@ -56,11 +56,18 @@ default_targets=(
     riscv64imac-unknown-none-elf
     riscv64gc-unknown-none-elf
 
-    # mips
+    # mips32r2
     mips-unknown-none # custom target
     mipsel-unknown-none
+    # mips32r6
+    mipsisa32r6-unknown-none   # custom target
+    mipsisa32r6el-unknown-none # custom target
+    # mips64r2
     mips64-unknown-none   # custom target
     mips64el-unknown-none # custom target
+    # mips64r6
+    mipsisa64r6-unknown-none   # custom target
+    mipsisa64r6el-unknown-none # custom target
 )
 
 x() {